test fix : GameRound 횟수  Validation
