+++ src/main/java/step4/dto/RacingGameConditionDTO.java
@@ -1,24 +1,26 @@
 package step4.dto;
 
+import step4.domain.GameRound;
+
 public class RacingGameConditionDTO {
     private String carNames;
-    private int tryCount;
+    private GameRound gameRound;
 
 
     private RacingGameConditionDTO(String carNames, int tryCount) {
         this.carNames = carNames;
-        this.tryCount = tryCount;
+        gameRound = GameRound.of(tryCount);
     }
 
-    public static RacingGameConditionDTO of(String carNames , int tryCount){
-        return new RacingGameConditionDTO(carNames , tryCount);
+    public static RacingGameConditionDTO of(String carNames, int tryCount) {
+        return new RacingGameConditionDTO(carNames, tryCount);
     }
 
     public String getCarNames() {
         return carNames;
     }
 
-    public int getTryCount() {
-        return tryCount;
+    public GameRound getGameRound() {
+        return gameRound;
     }
 }
